Move GraphQL inputs to their own module
