+++ app/graphql/mutations/create_photo.rb
@@ -2,7 +2,7 @@ module Mutations
   class CreatePhoto < BaseMutation
     null true
 
-    argument :photo, Types::PhotoInput, required: true
+    argument :photo, Types::Inputs::PhotoInput, required: true
 
     field :photo, Types::PhotoType, null: true
 
@@ -20,7 +20,7 @@ module Mutations
       if !file
         raise GraphQL::ExecutionError, 'Image must be provided as a "file" query parameter'
       end
-      
+
       file_ext = file.tempfile.path.split(".")[1]
       file_name = "#{SecureRandom.hex}.#{file_ext}"
 
@@ -43,7 +43,7 @@ module Mutations
       begin
         if photoable_type == 'User'
           photoable = User.find(photoable_id)
-          
+
           boot_unauthorized_user unless photoable.id == current_user.id
         elsif photoable_type == 'Dog'
           photoable = Dog.find(photoable_id)
