--- app/graphql/types/authentication_input.rb
@@ -1,7 +1,0 @@
-module Types
-  class AuthenticationInput < BaseInputObject
-    argument :first_name, String, required: true
-    argument :last_name, String, required: true
-    argument :email, String, required: true
-  end
-end
