+++ app/graphql/types/inputs/base_input_object.rb
@@ -1,0 +1,7 @@
+module Types
+  module Inputs
+    class BaseInputObject < GraphQL::Schema::InputObject
+      argument_class Types::BaseArgument
+    end
+  end
+end
