+++ app/graphql/types/inputs/user_input.rb
@@ -1,0 +1,10 @@
+module Types
+  module Inputs
+    class UserInput < BaseInputObject
+      argument :first_name, String, required: false
+      argument :last_name, String, required: false
+      argument :short_desc, String, required: false
+      argument :long_desc, String, required: false
+    end
+  end
+end
