--- app/graphql/types/location_input.rb
@@ -1,8 +1,0 @@
-module Types
-  class LocationInput < BaseInputObject
-    argument :street_address, String, required: false
-    argument :city, String, required: false
-    argument :state, String, required: false
-    argument :zip_code, String, required: true
-  end
-end
