--- app/graphql/types/photo_input.rb
@@ -1,7 +1,0 @@
-module Types
-  class PhotoInput < BaseInputObject
-    argument :photoable_type, String, required: true # TODO: use enum instead
-    argument :photoable_id, Integer, required: true
-    argument :caption, String, required: false
-  end
-end
