--- app/graphql/types/user_input.rb
@@ -1,8 +1,0 @@
-module Types
-  class UserInput < BaseInputObject
-    argument :first_name, String, required: false
-    argument :last_name, String, required: false
-    argument :short_desc, String, required: false
-    argument :long_desc, String, required: false
-  end
-end
